requirements.txt
asn1crypto==0.24.0
cffi==1.11.2
cryptography==2.1.4
Django==2.0.1
django-encrypted-model-fields==0.5.3
idna==2.6
pycparser==2.18
pytz==2017.3
six==1.11.0
